v1.0 of Table Serializer
Simple table-to-string and string-to-table tool for Lua. The purpose of this script is to allow the user a way to serialize a table into a string and deserialize it back into a table whenever needed. Used in ReaScripts, this way you can save tables as strings inside the Reaper-project files and extract them as tables.

Script is not thoroughly tested and may contain bugs.